Move Instance to Authentication to more accurately match JEP-303
